refactor: move cmd/ to internal/cli/ and relocate ParseUpstreamFlag
