+++ internal/integrate/integrate_test.go
@@ -65,35 +65,6 @@ func Test_resolveUpstreamURL(t *testing.T) {
 	})
 }
 
-func Test_ParseUpstreamFlag(t *testing.T) {
-	t.Run("url only", func(t *testing.T) {
-		spec, err := ParseUpstreamFlag("url=git@github.com:org/repo.git")
-		require.NoError(t, err)
-		assert.Equal(t, "git@github.com:org/repo.git", spec.URL)
-		assert.Equal(t, "", spec.Version)
-		assert.Equal(t, "", spec.Subpath)
-		assert.Equal(t, "", spec.Token)
-	})
-	t.Run("all keys", func(t *testing.T) {
-		spec, err := ParseUpstreamFlag("url=https://github.com/org/repo.git,version=main,subpath=infra,token=tok")
-		require.NoError(t, err)
-		assert.Equal(t, "https://github.com/org/repo.git", spec.URL)
-		assert.Equal(t, "main", spec.Version)
-		assert.Equal(t, "infra", spec.Subpath)
-		assert.Equal(t, "tok", spec.Token)
-	})
-	t.Run("missing url returns error", func(t *testing.T) {
-		_, err := ParseUpstreamFlag("version=main")
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "url")
-	})
-	t.Run("unknown key returns error", func(t *testing.T) {
-		_, err := ParseUpstreamFlag("url=git@github.com:org/repo.git,branch=main")
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "branch")
-	})
-}
-
 func Test_NormalizeUpstreamURL(t *testing.T) {
 	t.Run("SSH and HTTPS same repo match", func(t *testing.T) {
 		assert.Equal(t,
